Move prometheus to a different file
