+++ kubernetes/unicorn-raindrops.yml
@@ -40,13 +40,15 @@ spec:
             - name: POD_NAME
               value: 'myapp-app-server'
             - name: POD_NAMESPACE
-              value: 'pod161'
+              value: 'unicorn-raindrops'
             - name: PROJECT
               value: 'my-project'
             - name: RG_PROMETHEUS_ENABLED
               value: 'true'
             - name: RG_UNICORN
               value: 'true'
+            - name: RG_FREQUENCY
+              value: '5000'
           securityContext:
             runAsNonRoot: true
             readOnlyRootFilesystem: true
